Updates
Improved the hashparse functionality so that it checks for reused hashes and will also detail which accounts are DA if DA list is available.
DA list can be generated with get_desc parameter. Also added a parse_hashes parameter to parse hashes from an existing list i.e. generated offline etc. Also lsa_dump parameter has an option of wmi which uses wmiexec to do a procdump for parsing with mimikatz